requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
openai>=1.0.0
pyyaml>=6.0
requests>=2.31.0
fastmcp>=2.0.0
apscheduler>=3.10.0
python-dateutil>=2.8.0
